Django/latest_movie/requirements.txt
Django==3.2.10
djangorestframework==3.14.0
django-environ==0.11.2
django-storages==1.14.2
boto3==1.26.51
Pillow==8.4.0
bcrypt==4.1.2
psycopg2